+++ src/App.css
@@ -1,97 +1,6 @@
-* {
-    margin: 0;
-    padding: 0;
-}
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap');
 
 body {
-    margin-top: 20px;
-    font-family: sans-serif;
-}
-
-.input-form {
-    margin: 0 auto;
-    width: 400px;
-    padding: 30px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-h2 {
-    text-align: center;
-    padding-bottom: 20px;
-    text-transform: uppercase;
-    font-size: 22px;
-    font-weight: bold;
-    color: #474747;
-}
-
-input {
-    background-color: lightgrey;
-    border: 1px solid lightgrey;
-    padding: 20px;
-    margin-bottom: 20px;
-    border-radius: 4px;
-    color: #474747;
-    font-size: 16px;
-    outline: none;
-    box-sizing: border-box;
-    transition-duration: 0.5s;
-}
-
-input:focus {
-    border: 1px solid blueviolet;
-}
-
-input::placeholder {
-    color: #474747;
-    font-size: 16px;
-    font-weight: bold;
-}
-
-textarea {
-    background-color: lightgrey;
-    height: 100px;
-    border: 1px solid lightgrey;
-    padding: 20px;
-    margin-bottom: 20px;
-    resize: none;
-    border-radius: 4px;
-    font-family: sans-serif;
-    color: #474747;
-    font-size: 16px;
-    outline: none;
-    transition-duration: 0.5s;
-}
-
-textarea:focus {
-    border: 1px solid blueviolet;
-}
-
-textarea::placeholder {
-    color: #474747;
-    font-size: 16px;
-    font-weight: bold;
-}
-
-button {
-    background-color: blueviolet;
-    color: white;
-    border: none;
-    padding: 20px;
-    border-radius: 4px;
-    font-size: 16px;
-    font-weight: bold;
-    text-transform: uppercase;
-    transition-duration: 0.5s;
-}
-
-button:hover {
-    background-color: #2c1c4b;
-    cursor: pointer;
-}
-.wrapper {
-    width: 600px;
-    margin: 0 auto;
+ font-family: 'Roboto';
 }
