--- src/components/Message/Message.css
@@ -1,34 +1,0 @@
-.msg-box {
-    width: inherit;
-    padding: 20px 10px 0 10px;
-    display: flex;
-    flex-direction: column;
-    box-sizing: border-box;
-}
-
-.date {
-    font-size: 12px;
-    font-weight: bold;
-    color: blueviolet;
-    padding-bottom: 10px;
-    width: inherit;
-}
-
-.msg {
-    font-size: 16px;
-    color: #474747;
-    padding-left: 20px;
-    padding-bottom: 10px;
-    width: 100%;
-    box-sizing: border-box;
-
-}
-
-.author {
-    font-size: 12px;
-    font-weight: bold;
-    color: grey;
-    /*text-align: right;*/
-    padding-left: 20px;
-    width: inherit;
-}
