+++ src/components/Message/Message.js
@@ -1,5 +1,3 @@
-import './Message.css'
-
 const Message = (props) => {
     return (
         <div className='msg-box'>
