--- src/components/MessageList/index.js
@@ -1,2 +1,0 @@
-import MessageList from './MessageList';
-export default MessageList;
