+++ src/index.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+//import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-
-const myName = "Alexander";
+import 'typeface-roboto';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App name={myName} />
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
