+++ src/utils.js
@@ -1,0 +1,14 @@
+let getDate = () => {
+    const currentDate = new Date();
+    const options = {
+        year: 'numeric',
+        month: 'numeric',
+        day: 'numeric',
+        timezone: 'UTC',
+        hour: 'numeric',
+        minute: 'numeric',
+    };
+   return currentDate.toLocaleString("ru", options);
+}
+
+export default getDate;
